code/app/requirements.txt
#api
ujson==3.0.0
pydantic==1.6.1
python-jose[cryptography]==3.2.0

#ml
tensorflow==2.3.1
tensorflow_text==2.3.0
tensorflow_hub==0.10.0
#torch==1.5.1

#generic
python-decouple==3.3
requests==2.24.0